Update image-gallery.html
feat: use filenames as image captions in gallery

- Add data-caption attribute to gallery links
- Update JavaScript to use data-caption instead of title attribute

This change allows using more readable image filenames as captions in both the gallery view and lightbox, without requiring additional configuration. For example, "my-vacation-photo.jpg" will be displayed as "my vacation photo".

The order of images in the gallery is determined by the filename. To control the order, prefix filenames with numbers, e.g.:
- "01-first-image.jpg"
- "02-second-image.jpg"
- "03-third-image.jpg"