Fix "jumping" between different galleries
The issue is that the multiple inserted script tags share a common state
and so the `images` array needs to be separated. Therefore the window
object is used as a shared state and in that we reference the current
page and in the page the respective gallery that we want to show. It is
very hacky and not pretty at all and if you need something like it, you
should probably find a better solution than this…